requirements.txt
Authlib==0.14.3
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.1
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
os==3.6
pycparser==2.20
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1